refactor: cacheController.go was renamed to cacheCleaner.go the file was moved to cache package
